requirements.txt
flask==3.0.3
requests==2.31.0
python-dotenv==1.0.1
